Fix Warning and Add menu

+++ SDLTutorial/Barrier.cpp
@@ -1,6 +1,6 @@
 #include "Barrier.h"
 
-Barrier::Barrier(SDL_Renderer*& renderer,int x,int y):Paddle(renderer)
+Barrier::Barrier(SDL_Renderer*& renderer,float x,float y):Paddle(renderer)
 {
 	this->x = x;
 	this->y = y;
